% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{check_params_in_functions}
\alias{check_params_in_functions}
\title{Check That Parameters Are In Functions}
\usage{
check_params_in_functions(params, func_list)
}
\arguments{
\item{params}{Character vector with parameter names}

\item{func_list}{List of functions to check}
}
\description{
Test that certain parameters are arguments for certain functions via
\pkg{testthat} functions.
}
\examples{
MCHT:::check_params_in_functions(c("x"), list(mean))
}
