% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StartupMessage.R
\name{.onAttach}
\alias{.onAttach}
\title{Package Attach Hook Function}
\usage{
.onAttach(lib, pkg)
}
\arguments{
\item{lib}{a character string giving the library directory where the package
defining the namespace was found}

\item{pkg}{a character string giving the name of the package}
}
\description{
Hook triggered when package attached
}
\examples{
MCHT:::.onAttach(.libPaths()[1], "MCHT")
}
