% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalFunctions.R
\name{gen_memo_rng}
\alias{gen_memo_rng}
\title{Memoised Random Variable Generation}
\usage{
gen_memo_rng(r, seed = NULL)
}
\arguments{
\item{r}{The random number generator}

\item{seed}{The seed value, to be passed to \code{\link[base]{set.seed}}}
}
\value{
A function that generates random numbers with a set seed and with
        memoization; accepts \code{seed} and all other arguments that could
        be passed to the original random number generator
}
\description{
Creates a function that generates random numbers with memoization
}
\details{
This is a function generator, with the returned function being one that can
handle a set seed and will remember if it needs to regenerate a new set of
random numbers. This allows both for control over random number generation
and for faster performance.
}
\examples{
memo_runif <- gen_memo_runif(runif)
memo_runif(10)
}
