% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{\%s\%}
\alias{\%s\%}
\title{Concatenate (With Space)}
\usage{
x \%s\% y
}
\arguments{
\item{x}{One object}

\item{y}{Another object}
}
\value{
A string combining \code{x} and \code{y} with a space separating them
}
\description{
Concatenate and form strings (with space separation)
}
\examples{
`\%s\%` <- CPAT:::`\%s\%`
"Hello" \%s\% "world"
}
