% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{\%s0\%}
\alias{\%s0\%}
\title{Concatenate (Without Space)}
\usage{
x \%s0\% y
}
\arguments{
\item{x}{One object}

\item{y}{Another object}
}
\value{
A string combining \code{x} and \code{y}
}
\description{
Concatenate and form strings (no space separation)
}
\examples{
`\%s0\%` <- CPAT:::`\%s0\%`
"Hello" \%s0\% "world"
}
