% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StartupMessage.R
\name{MCHT_startup_message}
\alias{MCHT_startup_message}
\title{Create Package Startup Message}
\usage{
MCHT_startup_message()
}
\description{
Makes package startup message.
}
\examples{
MCHT:::MCHT_startup_message()
}
