% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarloTest.R
\name{print.MCHTest}
\alias{print.MCHTest}
\title{Print \code{MCHTest}-Class Object}
\usage{
\method{print}{MCHTest}(x, ...)
}
\arguments{
\item{x}{The \code{MCHTest}-class object}

\item{...}{Other arguments, such as \code{prefix} (a string wrapped around
the first line; by default, \code{"\t"})}
}
\description{
Print an \code{link{MCHTest}}-class object.
}
\examples{
f <- MCHTest(mean, mean, seed = 100)
print(f)
}
