% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarloTest.R
\name{print.MCHTest}
\alias{print.MCHTest}
\title{Print \code{MCHTest}-Class Object}
\usage{
\method{print}{MCHTest}(f, prefix = "\\t")
}
\arguments{
\item{f}{The \code{MCHTest}-class object}
}
\description{
Print an \code{link{MCHTest}}-class object.
}
\examples{
MCHT:::print.MCHTest()  # TODO: EXAMPLE
}
